+++ src/app/activities/[id]/page.tsx
@@ -1,78 +1,9 @@
-'use client';
-
-import ExperienceImageViewer from './components/experience/ExperienceImageViewer';
-import ExperienceSummarySection from './components/experience/ExperienceSummarySection';
-import Divider from './components/common/Divider';
-import SectionTitle from './components/common/SectionTitle';
-import ExperienceLocationMap from './components/experience/ExperienceLocationMap';
-import ReservationBox from './components/reservation/ReservationBox';
-import ExperienceDescription from './components/experience/ExperienceDescription';
-import ReviewSection from './components/review/ReviewSection';
-import { dummyReviews } from './data/dummyReviews';
-
-const dummyExperience = {
-  id: 7,
-  title: '함께 배우면 즐거운 스트릿댄스',
-  description:
-    '안녕하세요! 저희 스트릿 댄스 체험을 소개합니다. 저희는 신나고 재미있는 스트릿 댄스 스타일을 가르칩니다. 크럼프는 세계적으로 인기 있는 댄스 스타일로, 어디서든 춤출 수 있습니다. 저희 체험에서는 새로운 스타일을 접할 수 있고, 즐거운 시간을 보낼 수 있습니다. 저희는 초보자부터 전문가까지 어떤 수준의 춤추는 사람도 가르칠 수 있도록 준비해놓았습니다. 저희와 함께 즐길 수 있는 시간을 기대해주세요! 각종 음악에 적합한 스타일로, 저희는 크럼프 외에도 전통적인 스트릿 댄스 스타일과 최신 스트릿 댄스 스타일까지 가르칠 수 있습니다. 저희 체험에서는 전문가가 직접 강사로 참여하기 때문에, 저희가 제공하는 코스는 어떤 수준의 춤추는 사람도 쉽게 이해할 수 있도록 준비해놓았습니다. 저희 체험을 참가하게 된다면, 즐거운 시간 뿐만 아니라 새로운 스타일을 접할 수 있을 것입니다.',
-  category: '투어',
-  price: 10000,
-  address: '서울특별시 강남구 테헤란로 427',
-  bannerImageUrl: '/assets/icons/test/img-main.jpg',
-  subImages: [
-    { id: 1, imageUrl: '/assets/icons/test/img-sub1.jpg' },
-    { id: 2, imageUrl: '/assets/icons/test/img-sub2.jpg' },
-    { id: 3, imageUrl: '/assets/icons/test/img-sub3.jpg' },
-  ],
-  totalCount: 1300,
-  averageRating: 4.74,
-};
+import ExperienceResponsiveLayout from './components/common/ExperienceResponsiveLayout';
 
 export default function ExperienceDetailPage() {
   return (
     <main className='min-h-screen pt-[56px] pb-[160px] px-[16px] md:px-[24px]'>
-      {/* 이미지 + 요약 */}
-      <div className='w-full lg:max-w-4xl lg:mx-auto'>
-        <ExperienceImageViewer
-          bannerImageUrl={dummyExperience.bannerImageUrl}
-          subImages={dummyExperience.subImages}
-        />
-        <ExperienceSummarySection
-          title={dummyExperience.title}
-          rating={dummyExperience.averageRating.toFixed(1)}
-          reviewCount={dummyExperience.totalCount}
-          address={dummyExperience.address}
-        />
-        <Divider />
-      </div>
-
-      {/* 위치 */}
-      <div className='w-full lg:max-w-4xl lg:mx-auto'>
-        <SectionTitle title='만나는 곳' subtitle={dummyExperience.address} />
-        <ExperienceLocationMap address={dummyExperience.address} />
-        <Divider />
-      </div>
-
-      {/* 설명 */}
-      <div className='w-full lg:max-w-4xl lg:mx-auto'>
-        <SectionTitle title='체험 설명' />
-        <ExperienceDescription description={dummyExperience.description} />
-        <Divider />
-      </div>
-
-      {/* 후기 */}
-      <div className='w-full lg:max-w-4xl lg:mx-auto'>
-        <SectionTitle title='후기' />
-        <ReviewSection
-          activityId={dummyExperience.id}
-          rating={dummyExperience.averageRating}
-          reviewCount={dummyExperience.totalCount}
-          reviews={dummyReviews}
-        />
-      </div>
-
-      {/* 예약 */}
-      <ReservationBox pricePerPerson={dummyExperience.price} />
+      <ExperienceResponsiveLayout />
     </main>
   );
 }
