+++ src/styles/reservation-calendar-picker.css
@@ -4,11 +4,13 @@
   font-family: 'Pretendard', sans-serif;
   font-size: 14px;
   line-height: 1.4;
-  color: #1f2937; /* gray-800 */
+  color: #1f2937;
 }
 
 /* 요일 (일~토) */
 .react-calendar__month-view__weekdays {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
   text-align: center;
   font-weight: 600;
   font-size: 1.6rem;
@@ -16,77 +18,27 @@
   margin-bottom: 8px;
 }
 
+.react-calendar__month-view__weekdays__weekday {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 36px;
+  font-weight: 500;
+  font-size: 1.4rem;
+  color: #333;
+}
+
 .react-calendar__month-view__weekdays abbr {
   text-decoration: none;
 }
 
-/* 날짜 타일 공통 */
-.react-calendar__tile {
-  height: 36px;
-  width: 36px; /* max-width 대신 width 사용 */
-  padding: 0; /* padding 제거 */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  border-radius: 0; /* 기본은 네모 */
-  background: transparent;
-  color: #111827; /* 검정 텍스트 */
-  font-size: 1.4rem;
-  font-weight: 400;
-  transition: none;
-  border: none;
+/* 일요일: 빨간색 */
+.react-calendar__month-view__weekdays__weekday:nth-child(1) abbr {
+  color: red;
 }
-
-/* hover 상태 */
-.react-calendar__tile:hover:not(:disabled) {
-  background: #f3f4f6;
-  color: #111827;
-}
-
-.react-calendar__tile--now {
-  border: 2px solid #ffa67a !important;
-  font-weight: 600;
-  border-radius: 50%; /* 100% 대신 50% 사용 */
-  box-sizing: border-box; /* 보더가 크기에 포함되도록 */
-}
-
-/* 선택된 날짜 - 가장 높은 우선순위로 설정 */
-.react-calendar__tile--active,
-.react-calendar__tile--active:enabled,
-.react-calendar__tile--active:hover {
-  background: #ff5500 !important;
-  color: white !important;
-  font-weight: 600;
-  border-radius: 100%;
-}
-
-/* 선택된 범위 일부 */
-.react-calendar__tile--hasActive {
-  background: #93c5fd !important;
-  color: white;
-  font-weight: 600;
-  border-radius: 100%;
-}
-
-/* 이웃 월 날짜 */
-.react-calendar__month-view__days__day--neighboringMonth {
-  color: #d1d5db !important;
-  background: transparent !important;
-  font-weight: normal;
-}
-
-/* 비활성 날짜 */
-.react-calendar__tile:disabled {
-  color: #d1d5db !important;
-  pointer-events: none;
-  background: transparent !important;
-}
-
-/* 비활성 날짜 hover 방지 */
-.react-calendar__tile:disabled:hover {
-  background: transparent !important;
-  color: #d1d5db !important;
+/* 토요일: 파란색 */
+.react-calendar__month-view__weekdays__weekday:nth-child(7) abbr {
+  color: blue;
 }
 
 /* 날짜 그리드 */
@@ -96,31 +48,84 @@
   gap: 6px;
 }
 
-/* 포커스 제거 */
-.react-calendar__tile:focus {
-  outline: none;
+/* 날짜 셀 */
+.react-calendar__tile {
+  width: 100%;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  background: transparent;
+  transition: none;
 }
 
-/* 선택된 날짜가 오늘과 같을 때 */
-.react-calendar__tile--now.react-calendar__tile--active {
+/* 날짜 텍스트 */
+.react-calendar__tile abbr {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 36px;
+  height: 36px;
+  border-radius: 50%;
+  font-size: 1.4rem;
+  font-weight: 400;
+  color: #111827;
+}
+
+/* hover 효과 */
+.react-calendar__tile:hover:not(:disabled) abbr {
+  background: #f3f4f6;
+}
+
+/* 오늘 날짜 (선택되지 않은 상태) */
+.react-calendar__tile--now abbr {
+  border: 2px solid #ffa67a;
+  font-weight: 600;
+  box-sizing: border-box;
+}
+
+/* 선택된 날짜 */
+.react-calendar__tile--active {
+  background: transparent !important;
+}
+.react-calendar__tile--active abbr {
+  background: #ff5500 !important;
+  color: white !important;
+  font-weight: 600;
+}
+
+/* 오늘 + 선택됨 */
+.react-calendar__tile--now.react-calendar__tile--active abbr {
   background: #3b82f6 !important;
   color: white !important;
 }
 
-/* 요일 전체 공통 스타일 */
-.react-calendar__month-view__weekdays__weekday {
-  text-align: center;
-  font-weight: 500;
-  font-size: 14px;
-  color: #333;
+/* 선택된 범위 일부 */
+.react-calendar__tile--hasActive abbr {
+  background: #93c5fd !important;
+  color: white;
+  font-weight: 600;
 }
 
-/* 일요일: 첫 번째 요일 */
-.react-calendar__month-view__weekdays__weekday:nth-child(1) abbr {
-  color: red;
+/* 이웃 월 날짜 */
+.react-calendar__month-view__days__day--neighboringMonth abbr {
+  color: #d1d5db !important;
+  background: transparent !important;
+  font-weight: normal;
 }
 
-/* 토요일: 마지막 요일 */
-.react-calendar__month-view__weekdays__weekday:nth-child(7) abbr {
-  color: blue;
+/* 비활성 날짜 */
+.react-calendar__tile:disabled {
+  background: transparent !important;
+}
+.react-calendar__tile:disabled abbr {
+  background: transparent !important;
+  color: #d1d5db !important;
+  pointer-events: none;
+}
+
+/* 포커스 제거 */
+.react-calendar__tile:focus {
+  outline: none;
 }
